Reorganizing update code and made this as an alternate to copying zis plugins
