+++ bin/commands/components/install/cli.ts
@@ -16,6 +16,6 @@ import * as configmgr from '../../../libs/configmgr';
 
 const upgrade = false; //would be true for upgrade.ts
 
-index.execute(std.getenv('ZWE_CLI_PARAMETER_COMPONENT_FILE'), std.getenv('ZWE_CLI_PARAMETER_AUTO_ENCODING'), (std.getenv('ZWE_CLI_PARAMETER_SKIP_ENABLE') === 'true'), std.getenv('ZWE_CLI_PARAMETER_HANDLER'), std.getenv('ZWE_CLI_PARAMETER_REGISTRY'), (std.getenv('ZWE_CLI_PARAMETER_DRY_RUN') === 'true'), upgrade, std.getenv('ZWE_CLI_PARAMETER_STEPLIB'));
+index.execute(std.getenv('ZWE_CLI_PARAMETER_COMPONENT_FILE'), std.getenv('ZWE_CLI_PARAMETER_AUTO_ENCODING'), (std.getenv('ZWE_CLI_PARAMETER_SKIP_ENABLE') === 'true'), std.getenv('ZWE_CLI_PARAMETER_HANDLER'), std.getenv('ZWE_CLI_PARAMETER_REGISTRY'), (std.getenv('ZWE_CLI_PARAMETER_DRY_RUN') === 'true'), upgrade, std.getenv('ZWE_CLI_PARAMETER_ZIS_PLUGIN_DATASETS') ? std.getenv('ZWE_CLI_PARAMETER_ZIS_PLUGIN_DATASETS').split(",") : undefined);
 
 configmgr.cleanupTempDir();
