+++ bin/commands/components/install/process-hook/cli.ts
@@ -11,4 +11,4 @@
 
 import * as std from 'cm_std';
 import * as index from './index';
-index.execute(std.getenv('ZWE_CLI_PARAMETER_COMPONENT_NAME'));
+index.execute(std.getenv('ZWE_CLI_PARAMETER_COMPONENT_NAME'), std.getenv('ZWE_CLI_PARAMETER_ZIS_PLUGIN_DATASETS') ? std.getenv('ZWE_CLI_PARAMETER_ZIS_PLUGIN_DATASETS').split(",") : undefined);
