+++ bin/commands/components/install/process-hook/index.ts
@@ -18,7 +18,7 @@ import * as component from '../../../../libs/component';
 import * as varlib from '../../../../libs/var';
 import { PathAPI as pathoid } from '../../../../libs/pathoid';
 
-export function execute(componentName: string) {
+export function execute(componentName: string, zisPluginDatasets: string[]) {
   common.requireZoweYaml();
   const ZOWE_CONFIG=config.getZoweConfig();
   // read extensionDirectory
@@ -58,5 +58,5 @@ export function execute(componentName: string) {
   }
 
   component.processZssPluginInstall(componentDir);
-  component.processZisPluginInstall(componentDir);
+  component.processZisPluginInstall(componentDir, zisPluginDatasets);
 }
